[#2212] CID25764:
Fixed errant logic to intended behavior

+++ iRODS/lib/core/src/rmUtil.cpp
@@ -12,20 +12,16 @@
 int
 rmUtil( rcComm_t *conn, rodsEnv *myRodsEnv, rodsArguments_t *myRodsArgs,
         rodsPathInp_t *rodsPathInp ) {
-    int i = 0;
-    int status = 0;
-    int savedStatus = 0;
-    collInp_t collInp;
-    dataObjInp_t dataObjInp;
-
-
     if ( rodsPathInp == NULL ) {
         return ( USER__NULL_INPUT_ERR );
     }
 
+    collInp_t collInp;
+    dataObjInp_t dataObjInp;
     initCondForRm( myRodsEnv, myRodsArgs, &dataObjInp, &collInp );
 
-    for ( i = 0; i < rodsPathInp->numSrc; i++ ) {
+    int savedStatus = 0;
+    for ( int i = 0; i < rodsPathInp->numSrc; i++ ) {
         if ( rodsPathInp->srcPath[i].objType == UNKNOWN_OBJ_T ) {
             getRodsObjType( conn, &rodsPathInp->srcPath[i] );
             if ( rodsPathInp->srcPath[i].objState == NOT_EXIST_ST ) {
@@ -37,6 +33,7 @@ rmUtil( rcComm_t *conn, rodsEnv *myRodsEnv, rodsArguments_t *myRodsArgs,
             }
         }
 
+        int status = 0;
         if ( rodsPathInp->srcPath[i].objType == DATA_OBJ_T ) {
             status = rmDataObjUtil( conn, rodsPathInp->srcPath[i].outPath,
                                     myRodsEnv, myRodsArgs, &dataObjInp );
@@ -53,22 +50,15 @@ rmUtil( rcComm_t *conn, rodsEnv *myRodsEnv, rodsArguments_t *myRodsArgs,
             return ( USER_INPUT_PATH_ERR );
         }
         /* XXXX may need to return a global status */
-        if ( status < 0 ) {
+        if ( status < 0 &&
+                status != CAT_NO_ROWS_FOUND ) {
             rodsLogError( LOG_ERROR, status,
                           "rmUtil: rm error for %s, status = %d",
                           rodsPathInp->srcPath[i].outPath, status );
             savedStatus = status;
         }
     }
-    if ( savedStatus < 0 ) {
-        return ( savedStatus );
-    }
-    else if ( status == CAT_NO_ROWS_FOUND ) {
-        return ( 0 );
-    }
-    else {
-        return ( status );
-    }
+    return savedStatus;
 }
 
 int
